THE ORBIT UPDATE
This update includes a lot of changes and is a result of a month’s
work. In order to break down the changes, they are listed by ‘page’:

/welcome
•	Welcome page overhaul. It now presents the new ‘Dozen Skulls’
exhibition and includes a unique photograph of the museum’s rare
thylacine skull.

/exhibitions
•	Collections renamed to exhibitions. Rather than categorising the
whole museum and splitting the specimens up into collections. The focus
has moved to a modular exhibition’s style, where new small enticing
exhibitions can be added, providing a “taster experience” of the museum.
•	A Dozen Skulls exhibition added. This is the first exhibition and it
contains a dozen interesting skull specimens ranging from a lion to a
flamingo. Royalty free images have been added, however they are subject
to change. Please note the information/text in each of the specimens is
a placeholder at the moment and due to space only the hedgehog has the
unity player present.
•	Fixed issues with the card view. The card style view listing the
specimens in an exhibition has been fixed. They no longer merge and the
footer is now clearly separate. Now the UI is responsive and is
supported on all devices, from phones to tablets to desktops.

/specimen
•	Specimen’s page information reordered. The order of the page has been
changed and the corresponding xml/php code has been edited to reflect
this. The order is now title, latin name, photograph in the wild, brief
introduction about animal, 3D Unity viewer, Unity viewer description,
CT scan Link. This new structure makes each specimen interesting
without seeming to overwhelming.
•	Unity Viewer fully optimised. After a lot of tinkering the balance
between speed and quality has been found. The textures are now only a
few KB with the rest of the file size due to the mesh of the object. It
works on all supported browsers, without input lag. The lighting has
been adapted so the planar view is more effectively highlighted.
•	Unity Player now has fullscreen view. Unity web player can now go
into fullscreen in all browsers without an issue and return to its
original size. However when running in Safari this function has been
disabled, as Safari does not allow for Keyboard input in fullscreen
view.
•	Unity Player controls explained. Now a little help popup box is
present for those new to exploring the specimen, teaching them the
controls of how to interact with it in 3D/planar view.
•	3D view instructions. Three levels of things to look out for in the
3D specimen have been included; varying from Easy, to Mild and finally
to Tough. With Easy aimed at Key Stage Level, Mild aimed at GCSE/A
Level, and Tough aimed at University Level.
•	DICOM Viewer still present in code but not displayed. Unity’s built
in planar view is better than the dmv web viewer, however it is less
accurate. In a future release the DICOM viewer will be implemented as a
link onto a separate page; due to the files being quite large. Another
bonus of the DICOM viewer is that it works on mobile devices whereas
Unity currently doesn’t.

/events
•	Page now includes embedded live twitter feed. The twitter feed
displays the latest things happening at the museum, and will be used to
advertise upcoming events.
•	Custom social links. The museum is gaining a larger online presence
and will be signing up to other social networks in the near future.

/general
•	Cross-browser support. Fully functioning on
Chrome/Firefox/Safari/Internet Explorer/Edge.
•	Full mobile support. The html/css has been adapted for mobiles
(through the use of @media). An important purpose of this is that the
website can be viewed by visitors of the museum on their phones. So
they can bring up detailed tailored information for each specimen when
they see it in person. The electronic exhibition works in harmony with
the physical counterpart.
•	Fixed all previous bugs. Stl viewer now changes with each different
specimens. Navbar-toggle has been fixed. PHP/XML now works flawlessly.
•	Favicon created. Museum now has a custom favicon and corresponding
titles for each page.